+++ Kommunicate/Classes/Views/KMPreChatUserFormView.swift
@@ -86,9 +86,9 @@ class KMPreChatUserFormView: UIView, Localizable {
 
     private func placeholderWith(text: String) -> NSAttributedString {
         return NSAttributedString(string: text, attributes: [
-            .foregroundColor: UIColor(173, green: 168, blue: 168),
+            .foregroundColor: UIColor(173, green: 168, blue: 168) ?? UIColor.lightGray,
             .font: UIFont(name: "HelveticaNeue-Medium", size: 16.0) ?? UIFont.systemFont(ofSize: 16.0)
-            ])
+        ])
     }
 
     private func setupLocalizedLabelTexts() {